Refactor: replies.tsx 타입 분리 및 img alt 태그 추가
