+++ src/app/[userNickname]/[listId]/_components/ListDetailOuter/Replies.tsx
@@ -4,16 +4,10 @@ import timeDiff from '@/lib/utils/timeDiff';
 import * as styles from './Replies.css';
 import Line from '/public/icons/horizontal_line.svg';
 import DeleteButton from '/public/icons/trash_can.svg';
-
-interface Reply {
-  userNickName: string;
-  userProfileImageUrl: string;
-  createdDate: string;
-  content: string;
-}
+import { ReplyType } from '../../mockData/mockdataType';
 
 interface RepliesProps {
-  replies: Reply[] | null | undefined;
+  replies: ReplyType[] | null | undefined;
 }
 
 function Replies({ replies }: RepliesProps) {
@@ -27,17 +21,17 @@ function Replies({ replies }: RepliesProps) {
     <>
       {replies && !showReplies && (
         <div className={styles.showMoreRepliesWrapper} onClick={handleShowReplies}>
-          <Line />
+          <Line alt="답글 구분선" />
           <div className={styles.showMoreReplies}>{`답글 ${replies?.length}개 더 보기`} </div>
         </div>
       )}
       {showReplies && (
         <ul className={styles.repliesWrapper}>
-          {replies?.map((reply: Reply, idx: number) => {
+          {replies?.map((reply: ReplyType, idx: number) => {
             return (
               <li key={idx.toString()} className={styles.repliesOuterWrapper}>
                 <div className={styles.replyWrapper} key={idx.toString()}>
-                  <img src={reply.userProfileImageUrl} className={styles.profileImage}></img>
+                  <img src={reply.userProfileImageUrl} className={styles.profileImage} alt="사용자 프로필 이미지"></img>
                   <div className={styles.replyContainer}>
                     <div className={styles.replyInformationWrapper}>
                       <span className={styles.replyWriter}>{reply.userNickName}</span>
@@ -46,7 +40,7 @@ function Replies({ replies }: RepliesProps) {
                     <div className={styles.replyContent}>{reply.content}</div>
                   </div>
                 </div>
-                <DeleteButton className={styles.deleteButton} />
+                <DeleteButton className={styles.deleteButton} alt="삭제 버튼" />
               </li>
             );
           })}
