+++ src/app/[userNickname]/[listId]/mockData/mockdataType.ts
@@ -1,0 +1,25 @@
+export interface RepliesType {
+  id: number;
+  userId: number;
+  userNickName: string;
+  userProfileImageUrl: string;
+  createdDate: string;
+  content: string;
+}
+
+export interface ReplyType {
+  userNickName: string;
+  userProfileImageUrl: string;
+  createdDate: string;
+  content: string;
+}
+
+export interface Comment {
+  id: number;
+  userId: number;
+  userName: string;
+  userProfileImageUrl: string;
+  createdDate: string;
+  content: string;
+  replies: RepliesType[] | null;
+}
